pyproject.toml
[tool.poetry]
name = "wb-products"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
redis = "^5.1.1"
celery = "^5.4.0"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
